Tf12

+++ testing/test_model.py
@@ -33,6 +33,7 @@ class SquareModel2(hb.model.Model):
     def setUp(self):
         self.p = hb.param.Variable([2,3], collections = ['global1', 'global2'])
         self.q = hb.param.Variable([2,3], collections = ['global2'])
+        self.v = hb.variationals.Gaussian([2,3], collections = ['global2'])
 
     @hb.model.AutoOptimize()
     def likelihood(self):
@@ -43,9 +44,11 @@ class test_model2(unittest.TestCase):
     def setUp(self):
         tf.set_random_seed(0)
         self.m = SquareModel2()
-        self.filename = 'saved_file'
-        if os.path.exists(self.filename):
-            os.remove(self.filename)
+        self.filename = './saved_file.dat'
+        self.files = [self.filename+'.data-00000-of-00001',
+        self.filename+'.index', self.filename+'.meta', 'checkpoint']
+        if os.path.exists(self.filename): os.remove(self.filename)
+        [os.remove(f) for f in self.files if os.path.exists(f)]
 
     def test_initialize_finalize(self):
         optimizer = self.m.likelihood()
@@ -74,7 +77,7 @@ class test_model2(unittest.TestCase):
         self.m.q = np.ones((2,3))
         self.m.save(self.filename)
         # make sure the file is generated
-        self.assertTrue(os.path.exists(self.filename))
+        self.assertTrue(os.path.exists(self.files[0]))
         # make sure the model.restore()
         self.m2 = SquareModel2()
         self.m2.restore(self.filename)
@@ -83,17 +86,27 @@ class test_model2(unittest.TestCase):
         # make sure initialize() does not change the value
         self.m2.initialize()
         self.assertTrue(np.allclose(self.m2.q.value, np.ones((2,3))))
-        # remove the generated files
-        if os.path.exists(self.filename):
-            os.remove(self.filename)
-        if os.path.exists(self.filename+str('.meta')):
-            os.remove(self.filename+str('.meta'))
-        if os.path.exists('checkpoint'):
-            os.remove('checkpoint')
+        self.setUp()
+
+    def test_parameterized(self):
+        """ make sure save method for parameterized works fine """
+        self.m.v.q_mu = np.ones((2*3))
+        self.m.v.save(self.filename)
+        # make sure the file is generated
+        self.assertTrue(os.path.exists(self.files[0]))
+        # make sure the model.restore() works
+        self.m2 = SquareModel2()
+        self.m2.v.restore(self.filename)
+        self.assertTrue(np.allclose(self.m2.v.q_mu.value, np.ones((2*3))))
+        self.assertFalse(np.allclose(self.m2.p.value, np.ones((2,3))))
+        # make sure initialize() does not change the value
+        self.m2.initialize()
+        self.assertTrue(np.allclose(self.m2.v.q_mu.value, np.ones((2*3))))
+        self.setUp()
 
 class MinibatchModel(hb.model.Model):
     def setUp(self):
-        self.d = hb.param.MinibatchData(np.random.randn(2,3,100))
+        self.d = hb.param.MinibatchData(np.random.randn(100,2,3))
         self.p = hb.param.Variable([2,3])
 
     @hb.model.AutoOptimize()
@@ -109,11 +122,17 @@ class test_model3(unittest.TestCase):
         self.m.likelihood().compile()
         self.assertTrue(self.m._index.data_size == 100)
         # change the data size
-        self.m.d = hb.param.MinibatchData(np.random.randn(2,3,200))
+        self.m.d = hb.param.MinibatchData(np.random.randn(200,2,3))
         self.assertTrue(self.m._index.data_size == 100)
         # compile
         self.m.likelihood().compile()
         self.assertTrue(self.m._index.data_size == 200)
+
+        # check the data size
+        test_feed = self.m.test_feed_dict(20)
+        with self.m.tf_mode():
+            d = self.m.run(self.m.d, feed_dict = test_feed)
+        self.assertTrue(d.shape == (20,2,3))
 
 class ModelWithKeyword(hb.model.Model):
     def setUp(self, key1, key2):
